Updated testing

+++ contracts/src/ISertnServiceManager.sol
@@ -32,14 +32,14 @@ interface ISertnServiceManagerErrors {
 
 interface ISertnServiceManagerEvents {
     event NewOperator(address opAddr_);
-    event NewModels(uint256[] modelId_);
+    event NewModels(uint96[] modelId_);
     event NewStrategies(address[] newSupportedTokens_);
     event NewTask(address indexed opAddr_, bytes indexed taskId_);
-    event TaskResponded(uint256 indexed model, bytes indexed taskId, ISertnServiceManagerTypes.TaskResponse task);
+    event TaskResponded(uint96 indexed model, bytes indexed taskId, ISertnServiceManagerTypes.TaskResponse task);
     event UpForSlashing(address indexed operator, bytes indexed taskId);
     event ProofRequested(address indexed operator, bytes indexed taskId);
     event OperatorSlashed(address indexed operator, bytes indexed taskId);
-    event ModelUpdated(uint256 indexed modelId, ISertnServiceManagerTypes.OperatorModel operatorModel);
+    event ModelUpdated(uint96 indexed modelId, ISertnServiceManagerTypes.OperatorModel operatorModel);
     event OpInfoChanged(address indexed _operator, bytes _opInfo);
     event OperatorDeleted(address indexed _operator, uint32[] opSetIds);
 }
@@ -54,20 +54,20 @@ interface ISertnServiceManagerTypes {
     }
 
     struct Operator {
-        uint256[] models_;
+        uint96[] models_;
         bytes32[] computeUnits_;
         bytes[] openTasks_;
         bytes[] submittedTasks_;
         bytes[] proofRequests_;
         uint256[] allocatedEth_;
         uint256 allocatedSer_;
-        uint32[2] proofRequestExponents_;
+        uint256[2] proofRequestExponents_;
         uint32 pausedBlock_;
     }
 
     struct OperatorModel {
         address operator_;
-        uint256 modelId_;
+        uint96 modelId_;
         uint32 maxBlocks_;
         IStrategy[] ethStrategies_;
         uint256[] ethShares_;
@@ -80,11 +80,11 @@ interface ISertnServiceManagerTypes {
     }
 
     struct Task {
-        uint256 operatorModelId_;
+        uint96 operatorModelId_;
         bytes inputs_;
         uint256 poc_;
         uint256 startTime_;
-        uint256 startingBlock_;
+        uint32 startingBlock_;
         bool proveOnResponse_;
         address user_;
     }
